+++ src/app/models/Product/index.js
@@ -4,8 +4,8 @@ const Schema = mongoose.Schema
 const Product = new Schema({
     product_name: String,
     id_product_type: String,
-    old_price: String,
-    new_price: String,
+    old_price: Number,
+    new_price: Number,
     product_image: String,
     product_image_name: String,
     product_status: String,
